+++ ControleAcesso/DTO/FiltroMovimentoDTO.cs
@@ -3,7 +3,6 @@
     public class FiltroMovimentoDTO
     {
         public ESentido Sentido { get; set; }
-        public ETipoMovimento TipoMovimento { get; set; }
-        public EStatusMovimento StatusMovimento { get;  set; }
+       public EStatusMovimento StatusMovimento { get;  set; }
     }
 }
