+++ ControleAcesso/DTO/MovimentosObservacaoDTO.cs
@@ -5,7 +5,6 @@ namespace ControleAcesso.Domain.DTO
     public class MovimentosObservacaoDTO
     {
         public Guid Id { get; set; }
-        public ETipoMovimento TipoMovimento { get;  set; }
         public string Observacao { get; set; }
